% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cardinal_func.R
\name{Cardinal_func}
\alias{Cardinal_func}
\title{Function calculating gamma value}
\usage{
Cardinal_func(
  XX_list,
  Par_list = NULL,
  Interaction = TRUE,
  log = FALSE,
  Acid_Param = NULL,
  Acid_Weights = TRUE
)
}
\arguments{
\item{XX_list:}{List with named elements with values to be evaluated}

\item{Par_list:}{List with elements (known parameters), same names as XX_list.\cr 
-If CM function is to be used names are: xmin,xmax,xopt and n (all scalars)\cr
-If SR function is to be used names are: alpha, mic\cr
 If Par_list is set to NULL the values in data "CardinalPar" is used.}

\item{Interaction:}{- If TRUE, interactions are calculated based on CM elements \cr}

\item{log:}{If the cardinal value is to be returned on log scale}

\item{Acid_Param:}{List with parameters for acids}

\item{Acid_Weights:}{Logical, if TRUE (default) acids are to be recalculated to 
consentrations}
}
\value{
A scalar >=0 and <=1.
}
\description{
Function calculating gamma value
}
\examples{
Load data

}
\references{
Augustin, J., Zuliani, V., Cornu, M., and Guillier, L. (2005). 
Growth rate and growth probability ofListeria monocytogenes in dairy, meat and seafood 
products in suboptimal conditions. Journal of Applied Microbiology, 99(5):1019-1042.
doi: https://doi.org/10.1111/j.1365-2672.2005.02710.x
}
\author{
Lars Erik Gangsei
}
