% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cardinal_func_integral.R
\name{Cardinal_func_integral}
\alias{Cardinal_func_integral}
\title{Function calculating gamma value over time}
\usage{
Cardinal_func_integral(
  Environment,
  pHpar,
  Par_list = NULL,
  Interaction = TRUE,
  log = FALSE
)
}
\arguments{
\item{Environment:}{List with the following elements (scalars)\cr
- t: Time\cr
- Temp: : one numeric value, to be passed to Constant_func() for temperature.\cr
- aW: one numeric value, to be passed to Constant_func() for aW.\cr
- Lactate: one numeric value, to be passed to Constant_func() for lactate.\cr
- Acetate: one numeric value, to be passed to Constant_func() for acetate.\cr
- deltaL: one numeric value, to be passed to Constant_func() for deltaL.\cr
- carb: one numeric value for the carbohydrate level}

\item{pHpar:}{List with necessary elements to be sent to pH_func()}

\item{Par_list:}{List with elements (known parameters), same names as XX_list.\cr 
-If CM function is to be used names are: xmin,xmax,xopt and n (all scalars)\cr
-If SR function is to be used names are: alpha, mic\cr
 If Par_list is set to NULL the values in data "CardinalPar" is used.}

\item{Interaction:}{- If TRUE, interactions are calculated based on CM elements}

\item{log:}{If the cardinal values are to be returned on log scale}
}
\value{
A vector of length n with elements >=0 and <=1, reflecting cardinal 
values at times defined by input argument t.
}
\description{
Function calculating gamma value over time
}
\examples{
Load data

}
\author{
Lars Erik Gangsei
}
