% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cardinal_func_time.R
\name{Cardinal_func_time}
\alias{Cardinal_func_time}
\title{Function calculating gamma value over time}
\usage{
Cardinal_func_time(
  t,
  Par_list = NULL,
  pHpar,
  carb = NULL,
  pH0 = NULL,
  Temppar,
  aWpar,
  lacpar,
  acepar,
  deltaLpar = 1,
  Interaction = TRUE,
  log = FALSE
)
}
\arguments{
\item{t:}{vector of lengt n with times (in days) for when to calculate cardinal value}

\item{Par_list:}{List with elements (known parameters), same names as XX_list.\cr 
-If CM function is to be used names are: xmin,xmax,xopt and n (all scalars)\cr
-If SR function is to be used names are: alpha, mic\cr
 If Par_list is set to NULL the values in data "CardinalPar" is used.}

\item{pHpar:}{List with necessary elements to be sent to pH_func()}

\item{Temppar:}{List with element beta0 to be passed to Constant_func()}

\item{aWpar:}{List with element beta0 to be passed to Constant_func()}

\item{lacpar:}{List with element beta0 to be passed to Constant_func()}

\item{acepar:}{List with element beta0 to be passed to Constant_func()}

\item{deltaLpar:}{List with element beta0 to be passed to Constant_func()}

\item{Interaction:}{- If TRUE, interactions are calculated based on CM elements}

\item{log:}{If the cardinal values are to be returned on log scale}
}
\value{
A vector of length n with elements >=0 and <=1, reflecting cardinal 
values at times defined by input argument t.
}
\description{
Function calculating gamma value over time
}
\examples{
Load data

}
\author{
Lars Erik Gangsei
}
