% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CM_func.R
\name{CM_func}
\alias{CM_func}
\title{Function calculating CM value}
\usage{
CM_func(x, Par)
}
\arguments{
\item{x:}{vector of length n. The environmental factor (typically temperature, pH or water activity)
to be evaluated.}

\item{Par:}{List with the following elements "xmin", "xmax" ,"xopt" and "n" (all scalars)}
}
\value{
a vector of length n with all elements in interval  >=0 and <=1.
}
\description{
Function calculating CM value
}
\examples{
# Load data

}
\references{
Augustin, J., Zuliani, V., Cornu, M., and Guillier, L. (2005). 
Growth rate and growth probability ofListeria monocytogenes in dairy, meat and seafood 
products in suboptimal conditions. Journal of Applied Microbiology, 99(5):1019-1042.
doi: https://doi.org/10.1111/j.1365-2672.2005.02710.x
}
\author{
Lars Erik Gangsei
}
