% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation_func.R
\name{Estimation_func}
\alias{Estimation_func}
\title{Functions for estimating pH and listeria concentrations}
\usage{
Estimation_func(
  InputListeria,
  beta2 = NULL,
  pHmin = 4.3,
  Par_List = NULL,
  Interaction = TRUE
)
}
\arguments{
\item{InputpH:}{NULL if pH model is known. If pH parameters (\eqn{\beta^2})
is to be estimated a data frame with 6 elements: \cr
- t: Times (in days) for observations.\cr
- Temp: Temperatures in Celsius\cr
- carb: \eqn{x_c}, carbohydrate level, see Gangsei et.a. for details.\cr
- pH: observed pH\cr
- pH0: Start pH.\cr
- pHmin: Minimum pH.\cr}

\item{Par_List_pH:}{List with the following parameters\cr
- beta2: \eqn{\beta^2}, see Gangsei et.a. for details.\cr
- Temppar: Cardinal parameters for temperature for "pH -bacterias".}

\item{InputListeria:}{data.frame (n x 8) with observations for Listeria containing
the following elements\cr
- t: time (in days)\cr
- Temperature: in Celsius\cr
- Acetate: Acetate level\cr
- Lactate: lactate level\cr
- aW: Wateractivity\cr
- carb: \eqn{x_c}, carbohydrate level, see Gangsei et.a. for details.\cr
- DeltaL: Free water in the matrix.\cr
- yt: Listeria density at times t (cfu's)\cr
- y0: Listerialevels at time 0
- pH0: Start pH.\cr
- pHmin: Minimum pH.\cr}

\item{Par_list:}{Named list which is passed to "Cardinal_func". If set to NULL
the values in data "CardinalPar" is used.}

\item{Interaction:}{Passed directly to "Cardinal_func". 
If TRUE, interactions are calculated based on CM elements.}
}
\value{
A list with 5 elements \cr
- beta2: Estimated value for \eqn{\beta^2}\cr
- mu_opt_w_D: Estimated value for \eqn{mu_{Opt}} based on weighted estimate and dynamic environment\cr
- mu_opt_f_D: Estimated value for \eqn{mu_{Opt}} based on unweighted estimate and dynamic environment\cr
- mu_opt_w_ND: Estimated value for \eqn{mu_{Opt}} based on weighted estimate and non-dynamic environment\cr
- mu_opt_f_ND: Estimated value for \eqn{mu_{Opt}} based on unweighted estimate and non-dynamic environment\cr
}
\description{
Functions for estimating pH and listeria concentrations
}
\author{
Lars Erik Gangsei
}
