% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interaction_func.R
\name{Interaction_func}
\alias{Interaction_func}
\title{Function calculating interaction between temperature, pH and wateractivity in cardinal modelling}
\usage{
Interaction_func(XX_list, Par_list)
}
\arguments{
\item{XX_list:}{Named list (normally "Temp", "pH", "aW","Lactat" and "Acetat"), each element containing one scalar 
with the value to be evaluated.}

\item{Par_list:}{Named list with same names as XX_list (normally "Temp", "pH", "aW", '"Lactat" and "Acetat") 
each list  contains elements (at least) "xmin", "xmax" and "xopt" (all scalars)}
}
\value{
A scalar >=0 and <=1.
}
\description{
Function calculating interaction between temperature, pH and wateractivity in cardinal modelling
}
\examples{
Load data

}
\references{
Augustin, J., Zuliani, V., Cornu, M., and Guillier, L. (2005). 
Growth rate and growth probability ofListeria monocytogenes in dairy, meat and seafood 
products in suboptimal conditions. Journal of Applied Microbiology, 99(5):1019-1042.
doi: https://doi.org/10.1111/j.1365-2672.2005.02710.x
}
\author{
Lars Erik Gangsei
}
