% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pH_func.R
\name{pH_func}
\alias{pH_func}
\title{Function for calculating pH-value}
\usage{
pH_func(t, pHpar)
}
\arguments{
\item{t:}{Vector of length n, representing times for when pH is evaluated.}

\item{pHpar:}{A list with the following scalar elements.\cr
- beta2 >= 0: pH change rate.\cr
- carb: carbohydrate level\cr
- temp: temperature\cr
- pH0: Start pH (pH at time 0).\cr
- pHmin: Cardinal value for minimum pH-value.\cr
- TempPar: A list with cardinal parameters for temperature. Passed to CM_func. 
If NULL Parameters from data CardinalPar is used.}
}
\value{
A vector of length n with pH observations.
}
\description{
Function for calculating pH-value
}
\examples{
# Load data

}
\references{
Used in Gangsei, ....
}
\author{
Lars Erik Gangsei
}
