% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_pH.R
\name{predict_pH}
\alias{predict_pH}
\title{Functions for simulating pH and listeria concentrations}
\usage{
predict_pH(t, Temp, carb, pH0, pH_min, beta2, Temppar = NULL)
}
\arguments{
\item{t:}{Times (in days) for observations, length m.}

\item{Temp:}{Temperatures in Celsius, length m (or scalar)}

\item{carb:}{Carbohydrate level,length m, or scalar}

\item{beta2:}{value for \eqn{\beta^2}}

\item{pH0:}{Start pH, length m or scalar}

\item{pH_min:}{Minimum pH, scalar}

\item{Temppar:}{Named list which is passed to "Cardinal_func" for temperature.}
}
\value{
predicted length m
}
\description{
Functions for simulating pH and listeria concentrations
}
\author{
Lars Erik Gangsei
}
